Budget App: Tests

+++ db/migrate/20231024145204_remove_user_id_and_add_index_to_author_id_in_categories.rb
@@ -1,6 +1,5 @@
 class RemoveUserIdAndAddIndexToAuthorIdInCategories < ActiveRecord::Migration[7.0]
   def change
-    remove_column :categories, :user_id
-    add_index :categories, :author_id
+    remove_column :categories, :user_id if column_exists?(:categories, :user_id)
   end
 end
